+++ src/lib/mrvPanels/mrvSettingsPanel.cpp
@@ -265,12 +265,92 @@ namespace mrv
 
             key = prefix + "File Sequences";
             value = settings->getValue<std::any>(key);
-            open = std_any_empty(value) ? 1 : std_any_cast<int>(value);
+            open = std_any_empty(value) ? 0 : std_any_cast<int>(value);
+            if (!open)
+                cg->close();
+
+
+            cg = new CollapsibleGroup(
+                g->x(), 110, g->w(), 20, "OTIO");
+            b = cg->button();
+            b->labelsize(14);
+            b->size(b->w(), 18);
+            b->callback(
+                [](Fl_Widget* w, void* d)
+                {
+                    CollapsibleGroup* cg = static_cast<CollapsibleGroup*>(d);
+                    if (cg->is_open())
+                        cg->close();
+                    else
+                        cg->open();
+
+                    const std::string& prefix = settingsPanel->tab_prefix();
+                    const std::string key = prefix + "OTIO";
+
+                    App* app = App::app;
+                    auto settings = app->settings();
+                    settings->setValue(key, static_cast<int>(cg->is_open()));
+
+                    settingsPanel->refresh();
+                },
+                cg);
+
+            cg->begin();
+
+            bg = new Fl_Group(g->x(), 230, g->w(), 24);
+            bg->box(FL_NO_BOX);
+            bg->begin();
+
+            mW = new Widget< Fl_Choice >(
+                g->x() + 130, 230, g->w() - 130, 20, _("Spatial Coordinates"));
+            m = mW;
+            m->labelsize(12);
+            m->align(FL_ALIGN_LEFT);
+            for (const auto& i : timeline::getSpatialLabels())
+            {
+                m->add(i.c_str());
+            }
+            m->copy_tooltip(_(R"(Use the Spatial Coordinates in OTIO files to position and size the clips.
+
+* None: Ignore the spatial coordinates.
+* Coordinates: Use the spatial coordinates when clips provide them.
+* Normalize: Use the spatial coordinates, and display clips without them at the size of the first clip.  Use this to play clips of differing resolutions at the same size.)"));
+            m->value(settings->getValue<int>("OTIO/Spatial"));
+
+            mW->callback(
+                [=](auto o)
+                {
+                    int v = o->value();
+                    settings->setValue("OTIO/Spatial", v);
+                    refresh_movie_cb(nullptr, p.ui);
+                });
+
+            bg->end();
+
+            auto cV = new Widget< Fl_Check_Button >(
+                g->x() + 90, 250, g->w(), 20,
+                _("Compatibility"));
+            c = cV;
+            c->labelsize(12);
+            c->value(settings->getValue<bool>("OTIO/Compatibility"));
+            cV->callback(
+                [=](auto w)
+                {
+                    int v = w->value();
+                    settings->setValue("OTIO/Compatibility", v);
+                    refresh_movie_cb(nullptr, p.ui);
+                });
+
+            cg->end();
+
+            key = prefix + "OTIO";
+            value = settings->getValue<std::any>(key);
+            open = std_any_empty(value) ? 0 : std_any_cast<int>(value);
             if (!open)
                 cg->close();
 
             cg =
-                new CollapsibleGroup(g->x(), 210, g->w(), 22 * 8, _("Performance"));
+                new CollapsibleGroup(g->x(), 210, g->w(), 22 * 7, _("Performance"));
             cg->spacing(2);
             b = cg->button();
             b->labelsize(14);
@@ -393,7 +473,7 @@ namespace mrv
 
             bg->end();
 
-            auto cV = new Widget< Fl_Check_Button >(
+            cV = new Widget< Fl_Check_Button >(
                 g->x() + 90, 398, g->w(), 20,
                 _("FFmpeg YUV to RGB conversion"));
             c = cV;